requirements.txt
streamlit
openai
groq
beautifulsoup4
requests
moviepy==1.0.3
gradio_client>=0.10.0
edge-tts
Pillow
numpy
python-dotenv
imageio-ffmpeg
huggingface_hub
feedparser